fix(materials): address PR review comments for supplier linking feature
- Replace MockSupplierListQueryRepository with UrlSupplierListQueryRepository in
  MaterialCreate.tsx, MaterialUpdate.tsx, and both Next.js page files
- Inline MaterialSuppliersForm content directly into MaterialFormView (delete separate file)
- Rename availableSuppliers prop to suppliers across MaterialFormView, screens, and handlers
