+++ libs/ui/src/presentation/components/materials/index.ts
@@ -2,4 +2,3 @@ export * from './MaterialDeleteAlert';
 export * from './MaterialList';
 export * from './MaterialListItem';
 export * from './MaterialFormView';
-export * from './MaterialSuppliersForm';
