+++ libs/ui/src/presentation/screens/MaterialCreateScreen.tsx
@@ -10,7 +10,7 @@ export type MaterialCreateScreenProps = {
   isSubmitting: boolean;
   onLogoutPress: () => void;
   serverError?: string;
-  availableSuppliers: Supplier[];
+  suppliers: Supplier[];
   isLoadingSuppliers?: boolean;
 };
 
@@ -28,7 +28,7 @@ export const MaterialCreateScreen = (props: MaterialCreateScreenProps) => {
           isSubmitDisabled={props.isSubmitDisabled}
           isSubmitting={props.isSubmitting}
           serverError={props.serverError}
-          availableSuppliers={props.availableSuppliers}
+          suppliers={props.suppliers}
           isLoadingSuppliers={props.isLoadingSuppliers}
         />
       </ScrollView>
