+++ libs/ui/src/presentation/screens/MaterialUpdateHandler.tsx
@@ -50,7 +50,7 @@ export const MaterialUpdateHandler = ({
           : undefined
       }
       onLogoutPress={() => authLogout.dispatch({ type: 'LOGOUT' })}
-      availableSuppliers={supplierList.state.suppliers}
+      suppliers={supplierList.state.suppliers}
       isLoadingSuppliers={
         supplierList.state.type === 'idle' ||
         supplierList.state.type === 'loading'
